app/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.2
google-cloud-storage==1.37.1
firebase-admin==6.3.0
